+++ CHANGELOG.md
@@ -35,6 +35,7 @@ Changes since v2.4
   /api/forms/[form-id]/edit to /api/forms/edit.
 - The page addresses are no more prefixed with `/#/`, so for example the address
   of the catalogue page was changed from `/#/catalogue` to `/catalogue` (#1258)
+- More consistent user attributes in (application) API
 
 ### Additions
 - New field types: description, option, multiselect